%                              ~*~ Mode: Rd ~*~
%###############################################################################
%##                               murine.Rd
%###############################################################################
%## Author          : Mizanur Khondoker, Chris Glasbey, Bruce Worton
%## Created On      : 2008-04-22 14:39
%## Last Modified By: Mizanur Khondoker
%## Last Modified On: 2008-06-16 22:12
%## Update Count    : 9
%###############################################################################
%##
%## Copyright (C) 2008 Mizanur Khondoker
%## Distributed under GPL 2 or later
%###############################################################################


\name{murine}
\docType{data}
\alias{murine}
\title{
A subset of murine macrophage gene expression data from Khondoker et al.(2006).
}
\description{This data set contains a subset of 1000 rows of murine macrophage array1 
data (Khondoker \emph{et al.}, 2006). The data set has 4 columns corresponding to 
4 laser scans of one channel of a  microarray scanned at four different
scanner settings. This data set is used in an example to illustrate the
use of \code{multiscan} package.}

\usage{data(murine)}

\format{A 1000 by 4 data frame.}


\references{
  Khondoker, M. R., Glasbey, C. A. and Worton, B. J. (2006).
  Statistical estimation of gene expression using multiple laser scans of microarrays.
  \emph{Bioinformatics} \bold{22}, 215--219.
}

\keyword{datasets}
